(NOT TESTED) Finished SEN0244 driver

+++ NUCLEO-H755ZI/ASGC_Farming_System/CM7/Core/Inc/SEN0169.h
@@ -33,7 +33,7 @@ THIS CONVERSION FACTOR IS DETERMINED BY:
 #define SEN0169_INTERCEPT_OFFSET	15.498
 
 
-typedef float SEN0169_pH_Data;
+typedef double SEN0169_pH_Data;
 
 struct pH_Moving_Avg {
 	SEN0169_pH_Data pH_Data;
